requirements.txt
openai
langchain
retry
streamlit
streamlit-audiorecorder
git+https://github.com/stlukey/whispercpp.py
gTTS
